#174 Refactored BDD Feature tests
