+++ src/test/java/io/dashboardhub/pipelinedashboard/definitions/GuestDefinitions.java
@@ -5,21 +5,23 @@ import cucumber.api.java.en.Then;
 import cucumber.api.java.en.When;
 import io.dashboardhub.pipelinedashboard.PipelinedashboardApplicationTests;
 import org.junit.Assert;
+import org.openqa.selenium.By;
+import org.openqa.selenium.WebElement;
 
 public class GuestDefinitions extends PipelinedashboardApplicationTests {
 
     @Given("^I am not logged in$")
     public void I_am_not_logged_in() {
-        driver.get(withBaseUrl("/logout"));
+        try {
+            driver.findElement(By.id("logout")).submit();
+        } catch (org.openqa.selenium.NoSuchElementException e) {
+        }
+        driver.get(withBaseUrl("/"));
+        Assert.assertEquals("Welcome Guest, please Login", driver.findElement(By.id("login")).getText());
     }
 
     @When("^I try to access a secure page$")
     public void I_try_to_access_a_secure_page() {
         driver.get(withBaseUrl("/project"));
-    }
-
-    @Then("^I get redirected to the login page$")
-    public void I_get_redirected_to_the_login_page() {
-        Assert.assertEquals(withBaseUrl("/login"), driver.getCurrentUrl());
     }
 }
