+++ src/test/resources/io/dashboardhub/pipelinedashboard/definitions/project/secure.feature
@@ -4,3 +4,5 @@ Feature: Secure Project page
     Given I am logged in
     When I try to access the project page
     Then I get the project page
+
+  # ToDo: can not edit someone elses project
